Refactor ConsolidateStructAssignmentsTransform to use composition
