+++ src/plugin/src/transforms/consolidate-struct-assignments.ts
@@ -5,6 +5,9 @@
 import { Core, type MutableGameMakerAstNode } from "@gml-modules/core";
 import { FunctionalParserTransform } from "./functional-transform.js";
 import { CommentTracker } from "./utils/comment-tracker.js";
+import { StructAssignmentMatcher } from "./utils/struct-assignment-matcher.js";
+import { PropertyBuilder } from "./utils/property-builder.js";
+import { AssignmentCommentHandler } from "./utils/assignment-comment-handler.js";
 
 type CommentTools = {
     addTrailingComment: (...args: Array<unknown>) => unknown;
@@ -32,25 +35,6 @@ function normalizeCommentTools(commentTools) {
 
     return commentTools;
 }
-
-const STRUCT_EXPRESSION = "StructExpression";
-const VARIABLE_DECLARATION = "VariableDeclaration";
-const ASSIGNMENT_EXPRESSION = "AssignmentExpression";
-const MEMBER_DOT_EXPRESSION = "MemberDotExpression";
-const MEMBER_INDEX_EXPRESSION = "MemberIndexExpression";
-const IDENTIFIER = "Identifier";
-const LITERAL = "Literal";
-
-const IDENTIFIER_SAFE_PATTERN = /^[A-Za-z_][A-Za-z0-9_]*$/;
-
-type AllowTrailingCommentsBetweenOptions = {
-    tracker: CommentTracker;
-    left: number | undefined;
-    right: number | undefined;
-    precedingStatement: MutableGameMakerAstNode | null;
-    precedingProperty: MutableGameMakerAstNode | null;
-    commentTools: CommentTools;
-};
 
 function snapshotComment(comment) {
     return {
@@ -96,8 +80,17 @@ function restoreComment(comment, snapshot) {
 }
 
 export class ConsolidateStructAssignmentsTransform extends FunctionalParserTransform<ConsolidateStructAssignmentsTransformOptions> {
+    private readonly matcher: StructAssignmentMatcher;
+    private readonly propertyBuilder: PropertyBuilder;
+    private readonly commentHandler: AssignmentCommentHandler;
+
     constructor() {
         super("consolidate-struct-assignments", {});
+        this.matcher = new StructAssignmentMatcher();
+        this.propertyBuilder = new PropertyBuilder(
+            this.matcher.isIdentifierSafe.bind(this.matcher)
+        );
+        this.commentHandler = new AssignmentCommentHandler();
     }
 
     protected execute(
@@ -164,7 +157,9 @@ export class ConsolidateStructAssignmentsTransform extends FunctionalParserTrans
         }
 
         for (let index = 0; index < statements.length; index++) {
-            const initializer = this.getStructInitializer(statements[index]);
+            const initializer = this.matcher.getStructInitializer(
+                statements[index]
+            );
             if (!initializer) {
                 continue;
             }
@@ -197,9 +192,9 @@ export class ConsolidateStructAssignmentsTransform extends FunctionalParserTrans
             // The collected properties are assigned to the struct node; cast to
             // `MutableGameMakerAstNode` to allow mutation in-place with correct
             // typing for downstream transforms.
-            (structNode as MutableGameMakerAstNode).properties =
+            (structNode).properties =
                 collected.properties;
-            (structNode as MutableGameMakerAstNode).hasTrailingComma =
+            (structNode).hasTrailingComma =
                 collected.shouldForceBreak;
 
             statements.splice(index + 1, collected.count);
@@ -224,10 +219,11 @@ export class ConsolidateStructAssignmentsTransform extends FunctionalParserTrans
 
         while (cursor < statements.length) {
             const statement = statements[cursor];
-            const assignmentDetails = this.getStructPropertyAssignmentDetails(
-                statement,
-                identifierName
-            );
+            const assignmentDetails =
+                this.matcher.getStructPropertyAssignmentDetails(
+                    statement,
+                    identifierName
+                );
             if (!assignmentDetails) {
                 break;
             }
@@ -239,14 +235,14 @@ export class ConsolidateStructAssignmentsTransform extends FunctionalParserTrans
             }
 
             if (
-                !this.allowTrailingCommentsBetween({
+                !this.commentHandler.allowTrailingCommentsBetween(
                     tracker,
-                    left: lastEnd,
-                    right: start,
-                    precedingStatement: previousStatement,
-                    precedingProperty: lastProperty,
+                    lastEnd,
+                    start,
+                    previousStatement,
+                    lastProperty,
                     commentTools
-                })
+                )
             ) {
                 break;
             }
@@ -256,7 +252,9 @@ export class ConsolidateStructAssignmentsTransform extends FunctionalParserTrans
             }
 
             const property =
-                this.buildPropertyFromAssignment(assignmentDetails);
+                this.propertyBuilder.buildPropertyFromAssignment(
+                    assignmentDetails
+                );
             if (!property) {
                 break;
             }
@@ -267,7 +265,10 @@ export class ConsolidateStructAssignmentsTransform extends FunctionalParserTrans
                 end,
                 nextStart ?? Number.POSITIVE_INFINITY,
                 (comment) =>
-                    this.isAttachableTrailingComment(comment, statement)
+                    this.commentHandler.isAttachableTrailingComment(
+                        comment,
+                        statement
+                    )
             );
 
             if (attachableComments.length > 0) {
@@ -332,14 +333,14 @@ export class ConsolidateStructAssignmentsTransform extends FunctionalParserTrans
             : Number.POSITIVE_INFINITY;
 
         if (
-            !this.allowTrailingCommentsBetween({
+            !this.commentHandler.allowTrailingCommentsBetween(
                 tracker,
-                left: lastEnd,
-                right: nextBoundary,
-                precedingStatement: previousStatement,
-                precedingProperty: lastProperty,
+                lastEnd,
+                nextBoundary,
+                previousStatement,
+                lastProperty,
                 commentTools
-            })
+            )
         ) {
             tracker.rollback();
             touchedComments.forEach(({ comment, snapshot }) =>
@@ -367,354 +368,6 @@ export class ConsolidateStructAssignmentsTransform extends FunctionalParserTrans
             count: properties.length,
             shouldForceBreak
         };
-    }
-
-    private getStructInitializer(statement) {
-        if (!Core.isNode(statement)) {
-            return null;
-        }
-
-        if (statement.type === VARIABLE_DECLARATION) {
-            const declarator = Core.getSingleVariableDeclarator(
-                statement
-            ) as MutableGameMakerAstNode | null;
-            if (!Core.isNode(declarator)) {
-                return null;
-            }
-            if (!Core.isIdentifierNode(declarator.id)) {
-                return null;
-            }
-
-            if (Core.getNodeType(declarator.init) !== STRUCT_EXPRESSION) {
-                return null;
-            }
-
-            if (Core.getNodeType(declarator.init) !== STRUCT_EXPRESSION) {
-                return null;
-            }
-
-            if (
-                Array.isArray((declarator.init as any).properties) &&
-                (declarator.init as any).properties.length > 0
-            ) {
-                return null;
-            }
-
-            return {
-                identifierName: declarator.id.name,
-                structNode: declarator.init
-            };
-        }
-
-        if (statement.type === ASSIGNMENT_EXPRESSION) {
-            if (statement.operator !== "=") {
-                return null;
-            }
-
-            if (!Core.isIdentifierNode(statement.left)) {
-                return null;
-            }
-
-            if (Core.getNodeType(statement.right) !== STRUCT_EXPRESSION) {
-                return null;
-            }
-
-            if (
-                Array.isArray((statement.right as any).properties) &&
-                (statement.right as any).properties.length > 0
-            ) {
-                return null;
-            }
-
-            return {
-                identifierName: statement.left.name,
-                structNode: statement.right
-            };
-        }
-
-        return null;
-    }
-
-    private isIdentifierRoot(node, identifierName) {
-        return Core.isIdentifierNode(node) && node.name === identifierName;
-    }
-
-    private buildPropertyFromAssignment(
-        assignmentDetails
-    ): MutableGameMakerAstNode | null {
-        if (!assignmentDetails) {
-            return null;
-        }
-
-        const { assignment, propertyAccess } = assignmentDetails;
-        if (
-            !Core.isNode(assignment) ||
-            assignment.type !== ASSIGNMENT_EXPRESSION
-        ) {
-            return null;
-        }
-        const assignmentNode = assignment as MutableGameMakerAstNode;
-
-        if (!propertyAccess) {
-            return null;
-        }
-
-        const propertyKey = this.getPropertyKeyInfo(
-            propertyAccess.propertyNode
-        );
-        const propertyName = this.buildPropertyNameNode(propertyKey);
-        if (!propertyName) {
-            return null;
-        }
-
-        return {
-            type: "Property",
-            name: propertyName,
-            value: assignmentNode.right,
-            start:
-                Core.cloneLocation(
-                    this.getPreferredLocation(
-                        propertyAccess.propertyStart,
-                        assignmentNode.start
-                    )
-                ) ?? null,
-            end:
-                Core.cloneLocation(
-                    this.getPreferredLocation(
-                        assignmentNode.right?.end,
-                        assignmentNode.end
-                    )
-                ) ?? null
-        } as unknown as MutableGameMakerAstNode;
-    }
-
-    private getStructPropertyAssignmentDetails(statement, identifierName) {
-        if (
-            !Core.isNode(statement) ||
-            statement.type !== ASSIGNMENT_EXPRESSION
-        ) {
-            return null;
-        }
-
-        if (statement.operator !== "=") {
-            return null;
-        }
-
-        const propertyAccess = this.getStructPropertyAccess(
-            statement.left,
-            identifierName
-        );
-        if (!propertyAccess) {
-            return null;
-        }
-
-        return {
-            assignment: statement as MutableGameMakerAstNode,
-            propertyAccess
-        };
-    }
-
-    private getStructPropertyAccess(left, identifierName) {
-        if (!Core.isNode(left)) {
-            return null;
-        }
-
-        if (!this.isIdentifierRoot(left.object, identifierName)) {
-            return null;
-        }
-
-        if (left.type === MEMBER_DOT_EXPRESSION && Core.isNode(left.property)) {
-            return {
-                propertyNode: left.property,
-                propertyStart: left.property?.start
-            };
-        }
-
-        if (left.type === MEMBER_INDEX_EXPRESSION) {
-            const propertyNode = Core.getSingleMemberIndexPropertyEntry(left);
-            if (!Core.isNode(propertyNode)) {
-                return null;
-            }
-
-            return {
-                propertyNode,
-                propertyStart: propertyNode?.start
-            };
-        }
-
-        return null;
-    }
-
-    private getPropertyKeyInfo(propertyNode) {
-        if (!Core.isNode(propertyNode)) {
-            return null;
-        }
-
-        if (Core.isIdentifierNode(propertyNode)) {
-            return {
-                identifierName: propertyNode.name,
-                raw: propertyNode.name,
-                start: propertyNode.start,
-                end: propertyNode.end
-            };
-        }
-
-        if (
-            Core.isLiteralNode(propertyNode) &&
-            typeof propertyNode.value === "string"
-        ) {
-            const unquoted = Core.stripStringQuotes(propertyNode.value as any);
-            return {
-                identifierName: unquoted,
-                raw: propertyNode.value,
-                start: propertyNode.start,
-                end: propertyNode.end
-            };
-        }
-
-        return null;
-    }
-
-    private buildPropertyNameNode(propertyKey) {
-        if (!propertyKey) {
-            return null;
-        }
-
-        const identifierName = propertyKey.identifierName;
-        if (identifierName && this.isIdentifierSafe(identifierName)) {
-            return {
-                type: IDENTIFIER,
-                name: identifierName,
-                start: Core.cloneLocation(propertyKey.start) ?? null,
-                end: Core.cloneLocation(propertyKey.end) ?? null
-            };
-        }
-
-        if (typeof propertyKey.raw === "string") {
-            return {
-                type: LITERAL,
-                value: propertyKey.raw,
-                start: Core.cloneLocation(propertyKey.start) ?? null,
-                end: Core.cloneLocation(propertyKey.end) ?? null
-            };
-        }
-
-        return null;
-    }
-
-    private allowTrailingCommentsBetween(
-        options: AllowTrailingCommentsBetweenOptions
-    ) {
-        const {
-            tracker,
-            left,
-            right,
-            precedingStatement,
-            precedingProperty,
-            commentTools
-        } = options;
-        const commentEntries = tracker.getEntriesBetween(left, right);
-        if (commentEntries.length === 0) {
-            return true;
-        }
-
-        if (!precedingStatement) {
-            return false;
-        }
-
-        const expectedLine = Core.getNodeEndLine(precedingStatement);
-        if (typeof expectedLine !== "number") {
-            return false;
-        }
-
-        if (
-            commentEntries.some(
-                ({ comment }) =>
-                    !this.isTrailingLineCommentOnLine(comment, expectedLine)
-            )
-        ) {
-            return false;
-        }
-
-        const commentTarget = precedingProperty
-            ? (precedingProperty.value ?? precedingProperty)
-            : null;
-        for (const { comment } of commentEntries) {
-            if (comment.leadingChar === ";") {
-                comment.leadingChar = ",";
-            }
-
-            if (commentTarget) {
-                // Preserve historical metadata so the comment remains discoverable
-                // and mark it as a trailing comment so Prettier keeps it attached.
-                comment.enclosingNode = commentTarget;
-                commentTools.addTrailingComment(commentTarget, comment);
-            }
-        }
-
-        if (commentTarget) {
-            precedingProperty._hasTrailingInlineComment = true;
-        }
-
-        tracker.consumeEntries(commentEntries);
-        return true;
-    }
-
-    private isTrailingLineCommentOnLine(comment, expectedLine) {
-        if (!Core.isLineComment(comment)) {
-            return false;
-        }
-
-        return Core.getNodeStartLine(comment) === expectedLine;
-    }
-
-    private getPreferredLocation(primary, fallback) {
-        if (Core.isNode(primary)) {
-            return primary;
-        }
-        if (Core.isNode(fallback)) {
-            return fallback;
-        }
-        return null;
-    }
-
-    private isAttachableTrailingComment(comment, statement) {
-        if (!Core.isLineComment(comment)) {
-            return false;
-        }
-
-        const commentStart = comment.start;
-        if (
-            !Core.isObjectLike(commentStart) ||
-            typeof commentStart.line !== "number"
-        ) {
-            return false;
-        }
-
-        const statementEndLine = Core.getNodeEndLine(statement);
-        if (typeof statementEndLine !== "number") {
-            return false;
-        }
-
-        if (commentStart.line !== statementEndLine) {
-            return false;
-        }
-
-        const commentStartIndex = Core.getNodeStartIndex(comment);
-        const statementEndIndex = Core.getNodeEndIndex(statement);
-        if (
-            typeof commentStartIndex === "number" &&
-            typeof statementEndIndex === "number" &&
-            commentStartIndex <= statementEndIndex
-        ) {
-            return false;
-        }
-
-        return true;
-    }
-
-    private isIdentifierSafe(name) {
-        return typeof name === "string" && IDENTIFIER_SAFE_PATTERN.test(name);
     }
 }
